pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "trackbert"
version = "0.2.3"
authors = [
  { name="Kumi Mitterer", email="trackbert@kumi.email" },
]
description = "Python application tracking your shipments"
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pykeydelivery",
    "postat",
    "glsapi",
    "fedextrack",
    "dhltrack",
    "sqlalchemy",
    "alembic",
    "python-dateutil",
    "tabulate",
]

[project.urls]
"Homepage" = "https://kumig.it/kumitterer/trackbert"
"Bug Tracker" = "https://kumig.it/kumitterer/trackbert/issues"

[project.scripts]
trackbert = "trackbert.__main__:main"